Move unit and slot structs to their own file
